TextFormをTextInputBoxとし、ユニット化
また本コンポーネントのような独立的なコンポーネントをuniqueコンポーネントとして分類を明確化
